Build out initial action
This is largely based on the setup in balto-rubocop. Since this (much
like balto-prettier) only runs a command and doesn't deal with check
statuses, it's quite simple. The install_gems file is the key piece from
balto-rubocop that tries to selectively install only the gems you'll
need to run `stree write`. This can be error-prone, but it's also a
really nice speed improvement. Let's see how it does!