Smart install gems by building a custom Gemfile
This is a copy/paste with slight tweaks from a similar branch on
balto-rubocop. The big idea is that in order to cover more edge cases
and keep this process of installing a subset of gems and running
commands with just those predictable, relying on a Gemfile as the source
of truth works quite well.